telegram_bot/requirements.txt
python-telegram-bot==21.10
Flask==3.1.0
gunicorn==21.2.0
aiohttp==3.11.12